move secrets_manager_utils
