+++ src/zenml/secrets_managers/local/local_secrets_manager.py
@@ -30,8 +30,8 @@ from zenml.logger import get_logger
 from zenml.secret import SecretSchemaClassRegistry
 from zenml.secret.base_secret import BaseSecretSchema
 from zenml.secrets_managers.base_secrets_manager import BaseSecretsManager
+from zenml.secrets_managers.utils import decode_secret_dict, encode_secret
 from zenml.utils import yaml_utils
-from zenml.utils.secrets_manager_utils import decode_secret_dict, encode_secret
 
 logger = get_logger(__name__)
